[refactor] 삭제시 User를 확인하도록 수정

+++ src/main/java/nextstep/qna/domain/Answer.java
@@ -1,7 +1,7 @@
 package nextstep.qna.domain;
 
+import nextstep.qna.CannotDeleteException;
 import nextstep.qna.NotFoundException;
-import nextstep.qna.UnAuthorizedException;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
@@ -29,7 +29,10 @@ public class Answer extends Content{
         this.question = question;
     }
 
-    public DeleteHistory delete() {
+    public DeleteHistory delete(NsUser writer) throws CannotDeleteException {
+        if (!isOwner(writer)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
         this.deleted = true;
         return new DeleteHistory(ContentType.ANSWER, id, writer, LocalDateTime.now());
     }
